Fix use of undefined in Decimal operations at start of game
Sometimes, at the start of the game, undefined values were being passed
into Decimal methods. This implicitly converts them to be 0 in the
current version of break_eternity.js, but that behaviour should not be
relied on as the TypeScript types for Decimal methods state that they do
not support undefined values.

As a result, this commit fixes many instances where, at the start of the
game, some multipliers are set to zero due to this issue.

This commit intentionally uses ?? instead of || as the author does not
know what types are used - if it's possible that the left hand side
could be a number (not a Decimal) then || shouldn't be used (if the left
hand side was zero).  However, if [old browser support] is required, ||
should be used instead of ?? (assuming the types are correct).

This commit also uses Decimal.dOne and Decimal.dZero when needed instead
of constructing new Decimal instances.

[old browser support]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Operators/Nullish_coalescing_operator#browser_compatibility